Speedups for reloading refs when there are lots of tags.
This change uses GCD to schedule certain git commands for
background execution. All background commands occur on the
queue returned by PBGetWorkQueue(), and the results are
executed on the main queue. This results in a dramatic
improvement in the performance of reloadRefs when there
are lots of tags.

More improvements are possible with this approach, this is
just something to get things rolling.